+++ bindings/python/Math/src/Module.cpp
@@ -10,6 +10,7 @@
 #include <BipedalLocomotion/bindings/Math/Constants.h>
 #include <BipedalLocomotion/bindings/Math/Module.h>
 #include <BipedalLocomotion/bindings/Math/SchmittTrigger.h>
+#include <BipedalLocomotion/bindings/Math/Wrench.h>
 
 namespace BipedalLocomotion
 {
@@ -23,6 +24,7 @@ void CreateModule(pybind11::module& module)
 
     CreateConstants(module);
     CreateSchmittTrigger(module);
+    CreateWrench<double>(module, "d");
 }
 } // namespace Math
 } // namespace bindings