+++ utilities/balancing-position-control/script/blf-balancing-position-control.py
@@ -8,7 +8,7 @@ import datetime
 
 import bipedal_locomotion_framework.bindings as blf
 import yarp
-import idyntree.swig as idyn
+import idyntree.bindings as idyn
 
 import manifpy as manif
 
@@ -16,44 +16,79 @@ import manifpy as manif
 from pathlib import Path
 import sys
 
-sys.path.extend([str(Path(__file__).parent.resolve() / ".." / "share" / "BipedalLocomotionFramework" / "python")])
+sys.path.extend(
+    [
+        str(
+            Path(__file__).parent.resolve()
+            / ".."
+            / "share"
+            / "BipedalLocomotionFramework"
+            / "python"
+        )
+    ]
+)
 
 from balancing_position_control.wbc import WBC
-from balancing_position_control.zmp import evaluate_local_zmp, evaluate_global_zmp
-
 from datetime import timedelta
 
-def build_remote_control_board_driver(param_handler: blf.parameters_handler.IParametersHandler, local_prefix: str):
+
+def build_remote_control_board_driver(
+    param_handler: blf.parameters_handler.IParametersHandler, local_prefix: str
+):
     param_handler.set_parameter_string("local_prefix", local_prefix)
     return blf.robot_interface.construct_remote_control_board_remapper(param_handler)
 
 
-def build_contact_wrench_driver(param_handler: blf.parameters_handler.IParametersHandler, local_prefix: str):
+def build_contact_wrench_driver(
+    param_handler: blf.parameters_handler.IParametersHandler, local_prefix: str
+):
     param_handler.set_parameter_string("local_prefix", local_prefix)
     return blf.robot_interface.construct_generic_sensor_client(param_handler)
 
 
-def build_contact_wrenches_driver(params_contact_wrenches: blf.parameters_handler.IParametersHandler,
-                                  local_prefix: str):
+def build_contact_wrenches_driver(
+    params_contact_wrenches: blf.parameters_handler.IParametersHandler,
+    local_prefix: str,
+):
     # build contact wrenches polydrivers
     contact_wrenches_drivers = dict()
     contact_wrenches_names = dict()
     contact_wrenches_names["left_foot"] = []
     contact_wrenches_names["right_foot"] = []
 
-    for wrench_name in params_contact_wrenches.get_parameter_vector_string("left_contact_wrenches_group"):
+    for wrench_name in params_contact_wrenches.get_parameter_vector_string(
+        "left_contact_wrenches_group"
+    ):
         contact_wrenches_drivers[wrench_name] = build_contact_wrench_driver(
-            params_contact_wrenches.get_group(wrench_name), local_prefix)
-        assert contact_wrenches_drivers[wrench_name].is_valid()
-        contact_wrenches_names["left_foot"].append(
-            params_contact_wrenches.get_group(wrench_name).get_parameter_string("description"))
+            params_contact_wrenches.get_group(wrench_name), local_prefix
+        )
+        if not contact_wrenches_drivers[wrench_name].is_valid():
+            raise RuntimeError(
+                "Unable to create the contact wrench driver for " + wrench_name
+            )
 
-    for wrench_name in params_contact_wrenches.get_parameter_vector_string("right_contact_wrenches_group"):
+        contact_wrenches_names["left_foot"].append(
+            params_contact_wrenches.get_group(wrench_name).get_parameter_string(
+                "description"
+            )
+        )
+
+    for wrench_name in params_contact_wrenches.get_parameter_vector_string(
+        "right_contact_wrenches_group"
+    ):
         contact_wrenches_drivers[wrench_name] = build_contact_wrench_driver(
-            params_contact_wrenches.get_group(wrench_name), local_prefix)
-        assert contact_wrenches_drivers[wrench_name].is_valid()
+            params_contact_wrenches.get_group(wrench_name), local_prefix
+        )
+        if not contact_wrenches_drivers[wrench_name].is_valid():
+            raise RuntimeError(
+                "Unable to create the contact wrench driver for " + wrench_name
+            )
+
         contact_wrenches_names["right_foot"].append(
-            params_contact_wrenches.get_group(wrench_name).get_parameter_string("description"))
+            params_contact_wrenches.get_group(wrench_name).get_parameter_string(
+                "description"
+            )
+        )
 
     return contact_wrenches_drivers, contact_wrenches_names
 
@@ -61,7 +96,9 @@ def build_contact_wrenches_driver(params_contact_wrenches: blf.parameters_handle
 def build_kin_dyn(param_handler):
     rf = yarp.ResourceFinder()
     robot_model_path = rf.findFile("model.urdf")
-    joint_list = param_handler.get_group("ROBOT_CONTROL").get_parameter_vector_string("joints_list")
+    joint_list = param_handler.get_group("ROBOT_CONTROL").get_parameter_vector_string(
+        "joints_list"
+    )
     ml = idyn.ModelLoader()
     ml.loadReducedModelFromFile(robot_model_path, joint_list)
 
@@ -74,7 +111,9 @@ def get_base_frame(base_frame: str, kindyn: idyn.KinDynComputations):
     frame_base_index = kindyn.model().getFrameIndex(base_frame)
     link_base_index = kindyn.model().getFrameLink(frame_base_index)
     base_frame_name = kindyn.model().getLinkName(link_base_index)
-    return base_frame_name, kindyn.getRelativeTransform(frame_base_index, link_base_index)
+    return base_frame_name, kindyn.getRelativeTransform(
+        frame_base_index, link_base_index
+    )
 
 
 def create_new_spline(knots_positions, motion_duration: timedelta, dt: timedelta):
@@ -89,13 +128,22 @@ def create_new_spline(knots_positions, motion_duration: timedelta, dt: timedelta
 def main():
     # Before everything let use the YarpSink for the logger and the YarpClock as clock. These are functionalities
     # exposed by blf.
-    assert blf.text_logging.LoggerBuilder.set_factory(blf.text_logging.YarpLoggerFactory('balancing-position-control'))
-    assert blf.system.ClockBuilder.set_factory(blf.system.YarpClockFactory())
+    if not blf.text_logging.LoggerBuilder.set_factory(
+        blf.text_logging.YarpLoggerFactory("balancing-position-control")
+    ):
+        raise RuntimeError("Unable to set the logger factory")
+    if not blf.system.ClockBuilder.set_factory(blf.system.YarpClockFactory()):
+        raise RuntimeError("Unable to set the clock factory")
 
     param_handler = blf.parameters_handler.YarpParametersHandler()
-    assert param_handler.set_from_filename("blf-balancing-position-control-options.ini")
+    if not param_handler.set_from_filename(
+        "blf-balancing-position-control-options.ini"
+    ):
+        raise RuntimeError("Unable to load the parameters")
 
-    contact_force_threshold = param_handler.get_parameter_float("contact_force_threshold")
+    contact_force_threshold = param_handler.get_parameter_float(
+        "contact_force_threshold"
+    )
 
     dt = param_handler.get_parameter_datetime("dt")
 
@@ -105,30 +153,40 @@ def main():
     # Create the polydrivers
     poly_drivers, contact_wrenches_names = build_contact_wrenches_driver(
         params_contact_wrenches=param_handler.get_group("CONTACT_WRENCHES"),
-        local_prefix="balancing_controller")
+        local_prefix="balancing_controller",
+    )
 
     poly_drivers["REMOTE_CONTROL_BOARD"] = build_remote_control_board_driver(
         param_handler=param_handler.get_group("ROBOT_CONTROL"),
-        local_prefix="balancing_controller")
-    assert poly_drivers["REMOTE_CONTROL_BOARD"].is_valid()
+        local_prefix="balancing_controller",
+    )
+    if not poly_drivers["REMOTE_CONTROL_BOARD"].is_valid():
+        raise RuntimeError("Unable to create the remote control board driver")
 
     # just to wait that everything is in place
     blf.log().info("Sleep for two seconds. Just to be sure the interfaces are on.")
     blf.clock().sleep_for(datetime.timedelta(seconds=2))
 
     robot_control = blf.robot_interface.YarpRobotControl()
-    assert robot_control.initialize(param_handler.get_group("ROBOT_CONTROL"))
-    assert robot_control.set_driver(poly_drivers["REMOTE_CONTROL_BOARD"].poly)
+    if not robot_control.initialize(param_handler.get_group("ROBOT_CONTROL")):
+        raise RuntimeError("Unable to initialize the robot control")
+    if not robot_control.set_driver(poly_drivers["REMOTE_CONTROL_BOARD"].poly):
+        raise RuntimeError("Unable to set the driver for the robot control")
 
     # Create the sensor bridge
     sensor_bridge = blf.robot_interface.YarpSensorBridge()
-    assert sensor_bridge.initialize(param_handler.get_group("SENSOR_BRIDGE"))
-    assert sensor_bridge.set_drivers_list(list(poly_drivers.values()))
+    if not sensor_bridge.initialize(param_handler.get_group("SENSOR_BRIDGE")):
+        raise RuntimeError("Unable to initialize the sensor bridge")
+    if not sensor_bridge.set_drivers_list(list(poly_drivers.values())):
+        raise RuntimeError("Unable to set the drivers for the sensor bridge")
 
     # set the kindyn computations with the robot state
-    assert sensor_bridge.advance()
-    are_joint_ok, joint_positions, _ = sensor_bridge.get_joint_positions()
-    assert are_joint_ok
+    if not sensor_bridge.advance():
+        raise RuntimeError("Unable to advance the sensor bridge")
+
+    are_joints_ok, joint_positions, _ = sensor_bridge.get_joint_positions()
+    if not are_joints_ok:
+        raise RuntimeError("Unable to get the joint positions")
 
     base_frame = param_handler.get_parameter_string("base_frame")
     base_link, frame_T_link = get_base_frame(base_frame, kindyn=kindyn)
@@ -136,27 +194,67 @@ def main():
     right_contact_frame = param_handler.get_parameter_string("right_contact_frame")
     left_contact_frame = param_handler.get_parameter_string("left_contact_frame")
 
-    assert kindyn.setFloatingBase(base_link)
-    gravity = [0., 0., -blf.math.StandardAccelerationOfGravitation]
+    if not kindyn.setFloatingBase(base_link):
+        raise RuntimeError("Unable to set the floating base")
+
+    # set the robot state for the kindyn computations object
+    gravity = [0.0, 0.0, -blf.math.StandardAccelerationOfGravitation]
     base_velocity = idyn.Twist()
     base_velocity.zero()
-    joint_velocity = joint_positions * 0
-    assert kindyn.setRobotState(frame_T_link, joint_positions, base_velocity, joint_velocity, gravity)
+    desired_joint_velocities = joint_positions * 0
+    joint_velocities = desired_joint_velocities.copy()
+
+    if not kindyn.setRobotState(
+        frame_T_link, joint_positions, base_velocity, joint_velocities, gravity
+    ):
+        raise RuntimeError("Unable to set the robot state")
     initial_com_position = kindyn.getCenterOfMassPosition().toNumPy()
 
-    assert kindyn_with_measured.setFloatingBase(base_link)
-    assert kindyn_with_measured.setRobotState(frame_T_link, joint_positions, base_velocity, joint_velocity, gravity)
+    if not kindyn_with_measured.setFloatingBase(base_link):
+        raise RuntimeError("Unable to set the floating base")
+    if not kindyn_with_measured.setRobotState(
+        frame_T_link, joint_positions, base_velocity, joint_velocities, gravity
+    ):
+        raise RuntimeError("Unable to set the robot state")
+
+    global_cop_evaluator = blf.contacts.GlobalCoPEvaluator()
+    if not global_cop_evaluator.initialize(
+        param_handler.get_group("GLOBAL_COP_EVALUATOR")
+    ):
+        raise RuntimeError("Unable to initialize the global cop evaluator")
+
+    close_loop_with_zmp = param_handler.get_parameter_bool("close_loop_with_zmp")
+
+    # create zmp-com controller if close_loop_with_zmp is true
+    com_zmp_controller = blf.simplified_model_controllers.CoMZMPController()
+    if close_loop_with_zmp:
+        if not com_zmp_controller.initialize(
+            param_handler.get_group("COM_ZMP_CONTROLLER")
+        ):
+            raise RuntimeError("Unable to initialize the zmp-com controller")
 
     # create and initialize the IK
     ik = WBC(param_handler=param_handler.get_group("IK"), kindyn=kindyn)
     I_H_r_sole = kindyn.getWorldTransform(right_contact_frame)
     I_H_l_sole = kindyn.getWorldTransform(left_contact_frame)
-    assert ik.tasks["right_foot_task"].set_set_point(
-        blf.conversions.to_manif_pose(I_H_r_sole.getRotation().toNumPy(), I_H_r_sole.getPosition().toNumPy()))
-    assert ik.tasks["left_foot_task"].set_set_point(
-        blf.conversions.to_manif_pose(I_H_l_sole.getRotation().toNumPy(), I_H_l_sole.getPosition().toNumPy()))
-    assert ik.tasks["joint_regularization_task"].set_set_point(joint_positions)
-    assert ik.tasks["torso_task"].set_set_point(manif.SO3.Identity())
+    if not ik.tasks["right_foot_task"].set_set_point(
+        blf.conversions.to_manif_pose(
+            I_H_r_sole.getRotation().toNumPy(), I_H_r_sole.getPosition().toNumPy()
+        )
+    ):
+        raise RuntimeError("Unable to set the set point for the right foot task")
+    if not ik.tasks["left_foot_task"].set_set_point(
+        blf.conversions.to_manif_pose(
+            I_H_l_sole.getRotation().toNumPy(), I_H_l_sole.getPosition().toNumPy()
+        )
+    ):
+        raise RuntimeError("Unable to set the set point for the left foot task")
+    if not ik.tasks["joint_regularization_task"].set_set_point(joint_positions):
+        raise RuntimeError(
+            "Unable to set the set point for the joint regularization task"
+        )
+    if not ik.tasks["torso_task"].set_set_point(manif.SO3.Identity()):
+        raise RuntimeError("Unable to set the set point for the torso task")
 
     desired_joint_positions = joint_positions.copy()
 
@@ -167,76 +265,201 @@ def main():
     motion_timeout = param_handler.get_parameter_datetime("motion_timeout")
 
     spline = create_new_spline(
-        [initial_com_position + np.array([com_knots_delta_x[0], com_knots_delta_y[0], com_knots_delta_z[0]]),
-         initial_com_position + np.array([com_knots_delta_x[1], com_knots_delta_y[1], com_knots_delta_z[1]])],
-        motion_duration, dt)
+        [
+            initial_com_position
+            + np.array(
+                [com_knots_delta_x[0], com_knots_delta_y[0], com_knots_delta_z[0]]
+            ),
+            initial_com_position
+            + np.array(
+                [com_knots_delta_x[1], com_knots_delta_y[1], com_knots_delta_z[1]]
+            ),
+        ],
+        motion_duration,
+        dt,
+    )
 
     index = 0
     knot_index = 1
 
+    lipm_omega_square = (
+        blf.math.StandardAccelerationOfGravitation / initial_com_position[2]
+    )
+
+    desired_com_position = initial_com_position.copy()
+
     port = blf.yarp_utilities.BufferedPortVectorsCollection()
-    port.open("/balancing_controller/logger/data:o")
+    port.open("/balancing_position_controller/logger/data:o")
+
+    # switch to position direct
+    robot_control.set_control_mode(blf.robot_interface.YarpRobotControl.PositionDirect)
+
+    blf.log().info("Starting the balancing controller. Waiting for your input")
+    blf.log().info("Press enter to start the balancing controller")
+    input()
+    blf.log().info("Starting the balancing controller")
 
     while True:
-
         tic = blf.clock().now()
 
         # get the feedback
-        assert sensor_bridge.advance()
-        are_joint_ok, joint_positions, _ = sensor_bridge.get_joint_positions()
-        assert are_joint_ok
-        assert kindyn.setRobotState(frame_T_link, desired_joint_positions, base_velocity, joint_velocity, gravity)
-        assert kindyn_with_measured.setRobotState(frame_T_link, joint_positions, base_velocity, joint_velocity, gravity)
+        if not sensor_bridge.advance():
+            raise RuntimeError("Unable to advance the sensor bridge")
+        are_joints_ok, joint_positions, _ = sensor_bridge.get_joint_positions()
+        if not are_joints_ok:
+            raise RuntimeError("Unable to get the joint positions")
 
-        # solve the IK
-        com_spline_output = spline.get_output()
-        assert ik.tasks["com_task"].set_set_point(com_spline_output.position, com_spline_output.velocity)
-        assert ik.solver.advance()
-        assert ik.solver.is_output_valid()
+        are_joints_ok, joint_velocities, _ = sensor_bridge.get_joint_velocities()
+        if not are_joints_ok:
+            raise RuntimeError("Unable to get the joint velocities")
 
-        # integrate the system
-        desired_joint_positions += ik.solver.get_output().joint_velocity * dt.total_seconds()
+        if not kindyn.setRobotState(
+            frame_T_link,
+            desired_joint_positions,
+            base_velocity,
+            desired_joint_velocities,
+            gravity,
+        ):
+            raise RuntimeError("Unable to set the robot state")
+        if not kindyn_with_measured.setRobotState(
+            frame_T_link, joint_positions, base_velocity, joint_velocities, gravity
+        ):
+            raise RuntimeError("Unable to set the robot state")
 
-        # send the joint pose
-        assert robot_control.set_references(desired_joint_positions,
-                                            blf.robot_interface.YarpRobotControl.PositionDirect)
-
-        # send the data
         left_wrench = np.zeros(6)
         for cartesian_wrench_name in contact_wrenches_names["left_foot"]:
-            is_ok, wrench, _ = sensor_bridge.get_cartesian_wrench(cartesian_wrench_name)
-            assert is_ok
+            _, wrench, _ = sensor_bridge.get_cartesian_wrench(cartesian_wrench_name)
             left_wrench += wrench
 
         right_wrench = np.zeros(6)
         for cartesian_wrench_name in contact_wrenches_names["right_foot"]:
-            is_ok, wrench, _ = sensor_bridge.get_cartesian_wrench(cartesian_wrench_name)
-            assert is_ok
+            _, wrench, _ = sensor_bridge.get_cartesian_wrench(cartesian_wrench_name)
             right_wrench += wrench
 
-        global_zmp = evaluate_global_zmp(left_wrench=left_wrench, right_wrench=right_wrench,
-                                         l_sole_frame=left_contact_frame,
-                                         r_sole_frame=right_contact_frame,
-                                         contact_force_threshold=contact_force_threshold,
-                                         kindyn=kindyn)
-        global_zmp_from_measured = evaluate_global_zmp(left_wrench=left_wrench, right_wrench=right_wrench,
-                                                       l_sole_frame=left_contact_frame,
-                                                       r_sole_frame=right_contact_frame,
-                                                       contact_force_threshold=contact_force_threshold,
-                                                       kindyn=kindyn_with_measured)
-        local_zmp_left, _ = evaluate_local_zmp(wrench=left_wrench, contact_force_threshold=contact_force_threshold)
-        local_zmp_right, _ = evaluate_local_zmp(wrench=right_wrench, contact_force_threshold=contact_force_threshold)
+        # evaluate the global CoP using the desired joint state
+        left_contact = blf.contacts.ContactWrench()
+        left_contact.wrench = left_wrench
+        left_contact.pose = blf.conversions.to_manif_pose(
+            kindyn.getWorldTransform(left_contact_frame)
+        )
+        right_contact = blf.contacts.ContactWrench()
+        right_contact.wrench = right_wrench
+        right_contact.pose = blf.conversions.to_manif_pose(
+            kindyn.getWorldTransform(right_contact_frame)
+        )
+
+        if not global_cop_evaluator.set_input([left_contact, right_contact]):
+            raise RuntimeError("Unable to set the input for the global cop evaluator")
+        if not global_cop_evaluator.advance():
+            raise RuntimeError("Unable to advance the global cop evaluator")
+        global_zmp = global_cop_evaluator.get_output()
+
+        # evaluate the global CoP using the measured joint state
+        left_contact = blf.contacts.ContactWrench()
+        left_contact.wrench = left_wrench
+        left_contact.pose = blf.conversions.to_manif_pose(
+            kindyn_with_measured.getWorldTransform(left_contact_frame)
+        )
+        right_contact = blf.contacts.ContactWrench()
+        right_contact.wrench = right_wrench
+        right_contact.pose = blf.conversions.to_manif_pose(
+            kindyn_with_measured.getWorldTransform(right_contact_frame)
+        )
+        if not global_cop_evaluator.set_input([left_contact, right_contact]):
+            raise RuntimeError("Unable to set the input for the global cop evaluator")
+        if not global_cop_evaluator.advance():
+            raise RuntimeError("Unable to advance the global cop evaluator")
+        global_zmp_from_measured = global_cop_evaluator.get_output()
+
+        # use the CoM-ZMP controller
+        com_spline_output = spline.get_output()
+        # evaluate the desired ZMP using the LIP model
+        # ddx_com = omega^2 * (x_com - x_zmp)
+        desired_zmp = (
+            com_spline_output.position[:2]
+            - com_spline_output.acceleration[:2] / lipm_omega_square
+        )
+        desired_zmp = np.append(desired_zmp, 0.0)
+
+        # set the desired ZMP and the feedback if close_loop_with_zmp is true
+        if close_loop_with_zmp:
+            com_zmp_controller.set_set_point(
+                com_spline_output.velocity[:2],
+                com_spline_output.position[:2],
+                desired_zmp[:2],
+            )
+            com_zmp_controller.set_feedback(
+                kindyn_with_measured.getCenterOfMassPosition().toNumPy()[:2],
+                global_zmp_from_measured[:2],
+                0,
+            )
+            if not com_zmp_controller.advance():
+                raise RuntimeError("Unable to advance the CoM-ZMP controller")
+
+        # evaluate the desired CoM position
+        if close_loop_with_zmp:
+            desired_com_velocity = np.append(
+                com_zmp_controller.get_output(), com_spline_output.velocity[2]
+            )
+            desired_com_position[0:2] += (
+                com_zmp_controller.get_output() * dt.total_seconds()
+            )
+            desired_com_position[2] = com_spline_output.position[2]
+        else:
+            desired_com_velocity = com_spline_output.velocity
+            desired_com_position = com_spline_output.position
+
+        # solve the IK
+        if not ik.tasks["com_task"].set_set_point(
+            desired_com_position, desired_com_velocity
+        ):
+            raise RuntimeError("Unable to set the set point for the com task")
+        if not ik.solver.advance():
+            raise RuntimeError("Unable to advance the solver")
+        if not ik.solver.is_output_valid():
+            raise RuntimeError("The solver output is not valid")
+
+        # integrate the system
+        desired_joint_positions += (
+            ik.solver.get_output().joint_velocity * dt.total_seconds()
+        )
+        desired_joint_velocities = ik.solver.get_output().joint_velocity
+
+        # send the joint pose
+        if not robot_control.set_references(
+            desired_joint_positions, blf.robot_interface.YarpRobotControl.PositionDirect
+        ):
+            raise RuntimeError("Unable to set the references")
+
+        left_wrench = blf.math.Wrenchd(left_wrench)
+        local_zmp_left = left_wrench.get_local_cop()
+        right_wrench = blf.math.Wrenchd(right_wrench)
+        local_zmp_right = right_wrench.get_local_cop()
+
         com_from_desired = kindyn.getCenterOfMassPosition().toNumPy()
         com_from_measured = kindyn_with_measured.getCenterOfMassPosition().toNumPy()
 
         data = port.prepare()
-        data.vectors = {"global_zmp": global_zmp, "global_zmp_from_measured": global_zmp_from_measured,
-                        "local_zmp_left": local_zmp_left, "local_zmp_right": local_zmp_right,
-                        "com_from_desired": com_from_desired, "com_from_measured": com_from_measured}
+        data.vectors = {
+            "zmp::desired_planner": desired_zmp,
+            "zmp::measured::global::with_joint_desired": global_zmp,
+            "zmp::measured::global::with_joint_measured": global_zmp_from_measured,
+            "zmp::measured::local::left": local_zmp_left,
+            "zmp::measured::local::right": local_zmp_right,
+            "com::measured::with_joint_desired": com_from_desired,
+            "com::measured::with_joint_measured": com_from_measured,
+            "com::planned::position": com_spline_output.position,
+            "com::planned::velocity": com_spline_output.velocity,
+            "com::planned::acceleration": com_spline_output.acceleration,
+            "com::com_zmp::position": desired_com_position,
+            "com::com_zmp::velocity": desired_com_velocity,
+            "joints::desired::position": desired_joint_positions,
+        }
         port.write()
 
         # Final steps
-        assert spline.advance()
+        if not spline.advance():
+            raise RuntimeError("Unable to advance the spline")
 
         if index * dt >= motion_duration + motion_timeout:
             if knot_index + 1 >= len(com_knots_delta_x):
@@ -244,10 +467,27 @@ def main():
                 break
 
             spline = create_new_spline(
-                [initial_com_position + np.array(
-                    [com_knots_delta_x[knot_index], com_knots_delta_y[knot_index], com_knots_delta_z[knot_index]]),
-                 initial_com_position + np.array([com_knots_delta_x[knot_index + 1], com_knots_delta_y[knot_index + 1],
-                                                  com_knots_delta_z[knot_index + 1]])], motion_duration, dt)
+                [
+                    initial_com_position
+                    + np.array(
+                        [
+                            com_knots_delta_x[knot_index],
+                            com_knots_delta_y[knot_index],
+                            com_knots_delta_z[knot_index],
+                        ]
+                    ),
+                    initial_com_position
+                    + np.array(
+                        [
+                            com_knots_delta_x[knot_index + 1],
+                            com_knots_delta_y[knot_index + 1],
+                            com_knots_delta_z[knot_index + 1],
+                        ]
+                    ),
+                ],
+                motion_duration,
+                dt,
+            )
 
             knot_index += 1
             index = 0
@@ -262,6 +502,6 @@ def main():
     port.close()
 
 
-if __name__ == '__main__':
+if __name__ == "__main__":
     network = yarp.Network()
     main()